+++ docs/src/changes.md
@@ -1,5 +1,13 @@
 # Changes {#changes}
 
+## v1.5.0
+
+| Date       | Description                                  |
+|------------|----------------------------------------------|
+| 2025-05-07 | Issue-360: Fix python version deployment bug |
+| 2025-05-07 | Update minimum CMake version to 3.13         |
+| 2025-05-07 | Update GTest to v1.8.1                       |
+
 ## v1.4.0
 
 | Date       | Description                    |